Turn old "post scheduling service" into two stateless utility functions
*I recommend reviewing this diff with whitespace changes disabled.*

Previous commits created a post scheduler service. This creates a
problem, because there was also something called the "post scheduling
service" -- confusing!

Luckily, the latter is really just a pair of stateless functions, not a
stateful service. Let's drop the "service" moniker, which sidesteps the
problem.
